source/requirements.txt
crispy-bootstrap5==2024.2
django-crispy-forms==2.1
oauthlib==3.2.0
requests==2.25.1
pandas==2.2.2
numpy==1.26.4
tzdata==2024.1
googletrans==3.0.0
matplotlib==3.5.1
requests-oauthlib==2.0.0
matplotlib-inline==0.1.3